Initial work towards survey save feature.
This commit adds the file helpers for iOS, Android, and UWP to support saving the JSON survey responses to the local file system. It also introduces a "My Surveys" view to see all currently saved surveys. Currently, there is a context menu with two items, `Delete` and `Upload`.  Delete is wired up correctly while Upload still needs to be hooked up to the cloud service.