Fix Survey Question Title
Currently, the survey title displays "Question X out of Y", but the Y we're using is question count, which doesn't make a lot of sense as some questions can be skipped and the final question number doesn't line up with the total question count.

Instead, we're just displaying the current question number and for UWP, always just displaying "Survey Questions".

Fixes #60