Adds comments to view models and fixes up some logic for SurveyViewModel.
The previous question logic for survey view model was based on QuestionIDs, which are not necessarily guaranteed to be always in increasing order. Instead, we look for the question that comes previous from the current question number, so long as the user has previously provided an answer for that question.